+++ R/MLA.R
@@ -15,8 +15,7 @@
 #' set.seed(1)
 #' X <- rnorm(n = 100, mean = 0, sd = 10)
 #' Y <- X + rnorm(n=100, mean = 0, sd = 3)
-#' df <- data.frame(obs = X, pred = Y)
-#' MLA(df, obs = X, pred = Y)
+#' MLA(obs = X, pred = Y)
 #' }
 #' @rdname MLA
 #' @importFrom rlang eval_tidy quo