+++ R/SMAPE.R
@@ -12,7 +12,7 @@
 #' set.seed(1)
 #' X <- rnorm(n = 100, mean = 0, sd = 10)
 #' Y <- X + rnorm(n=100, mean = 0, sd = 3)
-#' MAPE(obs = X, pred = Y)
+#' SMAPE(obs = X, pred = Y)
 #' }
 #' @rdname SMAPE
 #' @importFrom rlang eval_tidy quo