+++ R/scatter_plot.R
@@ -1,4 +1,4 @@
-#' @title scatter.plot
+#' @title scatter_plot
 #' @description Scatter plot of Predictions and Observations.
 #' @param data (Optional) argument to call an existing data frame with the data. 
 #' @param obs Vector with observed values (numeric).
@@ -6,8 +6,8 @@
 #' @param orientation Argument of class string specifying the axis
 #' orientation, PO for predicted vs observed, and OP for
 #' observed vs predicted. Default is orientation = "PO".
-#' @param print.metrics boolean TRUE/FALSE to embed metrics in the plot. Default is FALSE.
-#' @param metrics.list vector or list of selected metrics to print on the plot.
+#' @param print_metrics boolean TRUE/FALSE to embed metrics in the plot. Default is FALSE.
+#' @param metrics_list vector or list of selected metrics to print on the plot.
 #' @param na.rm Logic argument to remove rows with missing values 
 #' (NA). Default is na.rm = TRUE.
 #' @return Object of class `ggplot`.
@@ -16,27 +16,27 @@
 #' \dontrun{
 #' X <- rnorm(n = 100, mean = 0, sd = 10)
 #' Y <- rnorm(n = 100, mean = 0, sd = 10)
-#' scatter.plot(obs = X, pre = Y)
+#' scatter_plot(obs = X, pre = Y)
 #' }
 #' @seealso 
 #'  \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{aes}}
-#' @rdname scatter.plot
+#' @rdname scatter_plot
 #' @importFrom ggplot2 ggplot geom_point aes geom_abline labs theme_bw
 #' theme
 #' @importFrom dplyr %>%
 #' @importFrom rlang eval_tidy quo
 #' @export
-scatter.plot <- function(data = NULL, 
+scatter_plot <- function(data = NULL, 
                          obs,
                          pred,
                          orientation = "PO",
-                         print.metrics = FALSE,
-                         metrics.list = NULL,
+                         print_metrics = FALSE,
+                         metrics_list = NULL,
                          na.rm = TRUE){
   
-  # STOP. Specify metrics.list
-  if (print.metrics == TRUE & is.null(metrics.list)) {
-    stop("Please, specify the 'metrics.list' arg. For example, metrics.list = c('R2','RMSE','NSE'). 
+  # STOP. Specify metrics_list
+  if (print_metrics == TRUE & is.null(metrics_list)) {
+    stop("Please, specify the 'metrics_list' arg. For example, metrics_list = c('R2','RMSE','NSE'). 
        Choose wisely, not more than 6 recommended to maintain an appropriate visualization") }
   
   # Get SMA slope and intercepts for printing
@@ -63,9 +63,9 @@ scatter.plot <- function(data = NULL,
   metrics.table <- rlang::eval_tidy(
     data = data,
     rlang::quo(
-    metrica::metrics.summary(data = data, obs = {{obs}}, pred = {{pred}}, 
+    metrica::metrics_summary(data = data, obs = {{obs}}, pred = {{pred}}, 
                              orientation = orientation,
-                    metrics.list = metrics.list) ) ) %>% 
+                    metrics_list = metrics_list) ) ) %>% 
     # Round numbers for clarity
     dplyr::mutate_if(base::is.numeric,~base::round(.,2)) 
   
@@ -137,7 +137,7 @@ scatter.plot <- function(data = NULL,
     )
   } #IF END
   
-  if (print.metrics == TRUE){
+  if (print_metrics == TRUE){
     plot <- 
       rlang::eval_tidy(data=data,
                        rlang::quo(