+++ R/uSD.R
@@ -10,7 +10,7 @@
 #' \dontrun{
 #' set.seed(1)
 #' X <- rnorm(n = 100, mean = 0, sd = 10)
-#' sdev.obs(obs = X)
+#' uSD(x = X)
 #' }
 #' @rdname uSD
 #' @importFrom rlang eval_tidy quo