+++ R/var_u.R
@@ -1,21 +1,21 @@
-#' @title var.u
-#' @description Uncorrected Sample Variance (var.u, not n-1).
+#' @title var_u
+#' @description Uncorrected Sample Variance (var_u, not n-1).
 #' @param data (Optional) argument to call an existing data frame containing the data.
 #' @param x Vector with numeric elements.
 #' @param na.rm Logic argument to remove rows with missing values 
 #' (NA). Default is na.rm = TRUE.
 #' @return Element of class `numeric`.
-#' @details Calculates the var.u of a vector x.
+#' @details Calculates the var_u of a vector x.
 #' @examples 
 #' \dontrun{
 #' set.seed(1)
 #' X <- rnorm(n = 100, mean = 0, sd = 10)
-#' var.u(x = X)
+#' var_u(x = X)
 #' }
-#' @rdname var.u
+#' @rdname var_u
 #' @importFrom rlang eval_tidy quo
 #' @export 
-var.u <- function(data=NULL,
+var_u <- function(data=NULL,
                   x,
                   na.rm = TRUE){
   result <- rlang::eval_tidy(